Move test bet functionality to dedicated Admin Test screen
Changes:
- Created new AdminTestScreen for test bet management
- Added isTestBet boolean field to Bet model schema
- Properly filters test bets (isTestBet=true) from real bets
- Reverted AdminDashboardScreen to payment management only
- Updated AccountScreen to use AdminTestScreen

Admin Test Screen Features:
- Create public test bets marked with isTestBet flag
- View and resolve test bets pending resolution
- All test bets properly isolated from real betting data
- Accessible via "Admin Testing Tools" menu option

Architecture:
- Separate concerns: AdminDashboardScreen = payments, AdminTestScreen = testing
- Test bets won't appear in regular bet lists (filtered by isTestBet=false)
- Clean separation between production and test data
